+++ src/constants/footer-social.ts
@@ -1,0 +1,28 @@
+export interface FooterSocialLink {
+	label: string
+	url: string
+	icon: string
+}
+
+export const FOOTER_SOCIAL_LINKS: FooterSocialLink[] = [
+	{
+		label: 'Instagram',
+		url: 'https://instagram.com/djmumbo',
+		icon: 'pi pi-instagram',
+	},
+	{
+		label: 'Twitter',
+		url: 'https://twitter.com/djmumbo',
+		icon: 'pi pi-twitter',
+	},
+	{
+		label: 'YouTube',
+		url: 'https://youtube.com/djmumbo',
+		icon: 'pi pi-youtube',
+	},
+	{
+		label: 'SoundCloud',
+		url: 'https://soundcloud.com/djmumbo',
+		icon: 'pi pi-cloud',
+	},
+]